localize test files
